Move export workflow into useProjectIO
